[cli] Fix error preventing velcom from launching
When installing the CLI without the --editable flag, the launch script failed to
properly import the main function. This commit simplifies the file structure by
moving main code to __init__.py and activates package auto-discovery for
setuptools so the command module is included when the package is installed.
