+++ cli/setup.cfg
@@ -3,7 +3,7 @@ name = velcom_cli
 version = 0.0.1
 
 [options]
-packages = velcom_cli
+packages = find:
 install-requires =
   requests >= 2.25.1
